% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{gh_app_set_commit_status}
\alias{gh_app_set_commit_status}
\title{Set Commit Status}
\usage{
gh_app_set_commit_status(repo, pkg, ref, buildlog, universe, jobsdata)
}
\arguments{
\item{repo}{full repo name for example "ropensci/magick"}

\item{pkg}{name of the package}

\item{ref}{hash of the commit to update}

\item{buildlog}{link to the build logs}

\item{universe}{name of the universe where packages were deployed to}

\item{jobsdata}{json string with jobs data}
}
\description{
Sets the commit status on a commit from a GitHub App.
Requires `GH_APP_ID` and `GH_APP_KEY` environment variables.
}
