% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment.R
\name{update_deployment_status}
\alias{update_deployment_status}
\title{Set Deployment Status}
\usage{
update_deployment_status(universe, package, ref, buildlog, source_status)
}
\arguments{
\item{universe}{name of the universe where packages were deployed to}

\item{package}{name of the package to be deployed}

\item{ref}{hash of the commit to update}

\item{buildlog}{link to the build logs}

\item{source_status}{string with result of building source pkg including vignettes}
}
\description{
Sets the deployment status in the monorepo
}
