% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{gh_app_set_commit_status}
\alias{gh_app_set_commit_status}
\title{Set Commit Status}
\usage{
gh_app_set_commit_status(repo, sha, state, url, context, description)
}
\arguments{
\item{repo}{full repo name for example "ropensci/magick"}

\item{sha}{hash of the commit to update}

\item{state}{one of 'pending', 'success', 'error', 'failure'}

\item{url}{link to the build logs}

\item{context}{which build this is for example "universe/docs"}

\item{description}{more information about this build}
}
\description{
Sets the commit status on a commit from a GitHub App.
Requires `GH_APP_ID` and `GH_APP_KEY` environment variables.
}
